backend/pyproject.toml
[project]
name = "WP-Project-Backend"
version = "0.1.0"
description = "Backend for Web Programming Project"
authors = [
    {name = "AliAlmasiZ", email = "ali1385almasi@gmail.com"},
    {name = "Parsa-1145"},
]
dependencies = [
    "django>=6.0",
    "django-rest-framework>=0.1.0", 
    "django-cors-headers>=4.9.0", 
    "drf-spectacular>=0.29.0", 
    "whitenoise>=6.11.0", 
    "dj-database-url>=3.1.0", 
    "dotenv>=0.9.9", 
    "djangorestframework-simplejwt>=5.5.1", 
    "Pillow>=12.1.1",
]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[project.optional-dependencies]
dev = [
    "ruff>=0.14.10",
    "black>=25.12.0",
    "pytest>=9.0.2",
    "django-pytest>=0.2.0",
]
[tool.pdm]
distribution = false

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "DJ", "DRF"]
ignore = []

[tool.black]
line-length = 88
target-version = ['py312']

[tool.setuptools]
packages = ["core", "accounts"]

